Add new TradingBot class that manages when each pari/strategy should run
Basically, the logic to track executions in done by the new dictionary: "last_executions".

In theory, I could have define that direct into the App.py. However, I can easily see us implementing new logic at the top of that. Specially logic related to "restart" the bot or "stop" the bot. We could potentially manage that using a API and Telegram (something for the future).
So, based on that, I believe it would be better encapsulating that in a new class. That's why I created this new  TradingBot class.