Add new app.py for managing the Backtest solution
The idea is to reorganize the app.py so that becomes more readable, clean and dry.
This way, when running a backtest, the user would manipulate only the values from the self.config. And that's it.

Previously, the user would need to manully comment some code lines and uncomment others. That was annoying and could be dangerous.

Note: Currently, is named App2.py because the  existing App.py is still in use.

squash