fix to show trend indicators value.
The method to get data from binance was being calculated always excluding the last candle (the active one). It has to be like that only for intraday data, so we don't get the current candle;
However, for trend data, we need to utilize the current candle, so we can correctly merge it with intraday dataset.