Add ability to include self.trend_class attribute dynamically
The self.trend_class is necessary only when running Backtest on Json_type.STRATEGY mode. When running on INTRADAY and TREND mode, that attribute is not used.

So, based on that, included a flag in the class constructor so that that attribute is set only when needed.
This would will make easier to run the Backtest in different scenarios.
squash