Add new temp JSON files to help the process of testing
There are so many classes for testing that  the process to manage all of those classes is becoming really hard.

So, based on that, created two temp JSON files so that we can list all of the available classes related to LONG and SHORT operations separately.

Notice that in the JSON for Short, the keys related to sell/buy are displayed in different order than was we have on JSON for Long. Ex: filter_sell_classes, trigger_sell_classes, etc.