Add ability to run SHORT operations even when running Strategies in the BacktestManagerIntraday
In the future, when running Strategies that should be managed by BacktestManagerStrategy (not BacktestManagerIntraday).
However, at this stage, we still use BacktestManagerIntraday.

I'm trying to analyze why the results of the backtesting for SHORT operation is not working as expected. So, I needed to create a new strategy with singles classes, and then I can set the "should_plot_chart" to true.