Fix baktesting not working when running Intraday mode for Short operations
Introduced by the commit that includes the logic to only open a new short position when trendAnalysis.is_downTrend.
Originally, trend_class is updated was updated to "UpTrend_alwaysTrend", but since we also support SHORT operations we would need to adjust that.

I considered including a logic to set  trend_class based on the given self.operation_type.  Either: DownTrend_AlwaysTrend or UpTrend_alwaysTrend. That would work well for Intraday mode. But I'm not sure if it would be good when running Strategy and Portfolio mode.

It's also important to mention that, ideally, BacktestManagerIntraday should not even have or use `trend_class`. That is something that should be handled by opther managers.
So, leave this as is for now.