Fix DB table created with the wrong name
I could just rename the table, but since we are kind a starting a new dashboard for backtest, I think it's fine starting the table from scratch.
So, also included the new `Side` column to be added as part of the table creation.
Note: I did some small adjustments in the SQL script, like:
	• created_at: to use 'with' time zone (keep consistent with the other datetime fields of this table - start_time/end_time)
	• best_indicators_combination: to use TEXT instead of varying(6000)

I also did some adjustments in other places that were using the wrong name.